+++ assets/javascripts/kitten/karl/carousel/flex-carousel.js
@@ -1,0 +1,27 @@
+import React from 'react'
+import Radium, { StyleRoot } from 'radium'
+import { FlexCarousel } from 'kitten/components/carousel/flex-carousel'
+import { Container } from 'kitten/components/grid/container'
+import { Marger } from 'kitten/components/layout/marger'
+import { KarlExampleTitle } from 'kitten/karl/examples/title'
+
+export const KarlFlexCarousel = () => (
+  <StyleRoot>
+    <KarlExampleTitle>Flex Carousel</KarlExampleTitle>
+    <Container>
+      <FlexCarousel
+        slideContent={({ content }) => {
+          return (
+            <FlexCarouselContent
+              title={content.title}
+              titleLineBreak={content.titleLineBreak}
+              paragraph={content.paragraph}
+              paragraphLineBreak={content.paragraphLineBreak}
+              button={content.button}
+            />
+          )
+        }}
+      />
+    </Container>
+  </StyleRoot>
+)