+++ lib/kitten/configuration.rb
@@ -275,10 +275,8 @@ module Kitten
           id: 'carousel',
           label: 'Carousel',
           components: [
-            {
-              path: 'components/carousel/carousel',
-              name: 'Carousel with ProjectCard'
-            },
+            { path: 'components/carousel/carousel', name: 'Carousel with ProjectCard' },
+            { path: 'components/carousel/flex-carousel', name: 'Flex Carousel'},
           ],
         },
       ]