+++ spec/dummy/client/entries/app-kitten.js
@@ -325,9 +325,8 @@ import {
 } from 'kitten/karl/typography/paragraph'
 
 // Carousel
-import {
-  KarlCarouselProjectCard,
-} from 'kitten/karl/carousel/carousel'
+import { KarlCarouselProjectCard } from 'kitten/karl/carousel/carousel'
+import { KarlFlexCarousel } from 'kitten/karl/carousel/flex-carousel'
 
 ReactOnRails.register({
   // Box
@@ -596,4 +595,5 @@ ReactOnRails.register({
 
   // Carousel
   KarlCarouselProjectCard,
+  KarlFlexCarousel,
 })